Update LemmyApi (#1486)
* Update LemmyApi

* Remove some remnants

* Remove some remnants

* Fix #1487

+++ app/src/main/java/com/jerboa/ui/components/home/HomeActivity.kt
@@ -289,7 +289,7 @@ fun MainPostListingsContent(
                     homeViewModel.likePost(
                         CreatePostLike(
                             post_id = postView.post.id,
-                            score = newVote(postView.my_vote, VoteType.Upvote).toLong(),
+                            score = newVote(postView.my_vote, VoteType.Upvote),
                         ),
                     )
                 }
@@ -305,7 +305,7 @@ fun MainPostListingsContent(
                     homeViewModel.likePost(
                         CreatePostLike(
                             post_id = postView.post.id,
-                            score = newVote(postView.my_vote, VoteType.Downvote).toLong(),
+                            score = newVote(postView.my_vote, VoteType.Downvote),
                         ),
                     )
                 }
